# openapi.yaml
openapi: 3.0.0
info:
  title: My Collection
  description: >-
    ### Welcome to Postman! This is your first collection.


    Collections are your starting point for building and testing APIs. You can
    use this one to:


    • Group related requests  

    • Test your API in real-world scenarios  

    • Document and share your requests


    Update the name and overview whenever you’re ready to make it yours.


    [Learn more about Postman
    Collections.](https://learning.postman.com/docs/collections/collections-overview/)
  version: 1.0.0
servers:
  - url: http://{{baseurl}}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
paths:
...
  /api/users/appeals:
    post:
      tags:
        - CSC309 Project > Users > Appeals
      summary: Create appeal
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                content: Test appeal Jaic
      security:
        - bearerAuth: []
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    get:
      tags:
        - CSC309 Project > Users > Appeals
      summary: Get appeals
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
          example: '1'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /api/users/edit:
    patch:
      tags:
        - CSC309 Project > Users
      summary: Edit User Info
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                avatarId: https://example.com/images/new-avatar.png
                favoriteTeamId: 64
      security:
        - bearerAuth: []
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /api/threads/{threadId}/posts:
    post:
      tags:
        - CSC309 Project > Forums > Posts > GET /api/posts/[id]/translate
      summary: Create a post in a different language
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                content: Bonjour, mon nom n'a pas d'importance.
      security:
        - bearerAuth: []
      parameters:
        - name: threadId
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    get:
      tags:
        - CSC309 Project > Forums > Posts > GET /api/threads/[id]/posts
      summary: Get Posts of a Thread
      security:
        - bearerAuth: []
      parameters:
        - name: threadId
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /api/threads/{threadId}/sentiment:
    get:
      tags:
        - CSC309 Project > Forums > Threads > GET /api/threads/[id]/sentiment
      summary: Get Sentiment
      security:
        - bearerAuth: []
      parameters:
        - name: threadId
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...